test: split sidebar actions fixture budget
